0425 11 20 : 파티디테일 현 파티원목록 수정 필요

+++ chamchi/src/main/java/com/ERR/infra/party/PartyDto.java
@@ -7,7 +7,7 @@ public class PartyDto {
 	private String partyName;
 	private String partyDept;
 	private Integer partyMatchTypeCd;
-	private int partyDelNy;
+	private Integer partyDelNy;
 	private Date partyRegDate;
 	private Date partyModDate;
 
@@ -18,144 +18,113 @@ public class PartyDto {
 	private String xMasterNy;
 	private String xPartyJoinNy;
 	private Integer xMemberTierCd;
-	private Integer xMemberVictoryRate;
-	private double xMemberKDA;
+	private Double xMemberVictoryRate;
+	private Double xMemberKDA;
 	private Date xPartyJoinDate;
-
+	
 	public String getPartySeq() {
 		return partySeq;
 	}
-
 	public void setPartySeq(String partySeq) {
 		this.partySeq = partySeq;
 	}
-
 	public String getPartyName() {
 		return partyName;
 	}
-
 	public void setPartyName(String partyName) {
 		this.partyName = partyName;
 	}
-
 	public String getPartyDept() {
 		return partyDept;
 	}
-
 	public void setPartyDept(String partyDept) {
 		this.partyDept = partyDept;
 	}
-
 	public Integer getPartyMatchTypeCd() {
 		return partyMatchTypeCd;
 	}
-
 	public void setPartyMatchTypeCd(Integer partyMatchTypeCd) {
 		this.partyMatchTypeCd = partyMatchTypeCd;
 	}
-
-	public int getPartyDelNy() {
+	public Integer getPartyDelNy() {
 		return partyDelNy;
 	}
-
-	public void setPartyDelNy(int partyDelNy) {
+	public void setPartyDelNy(Integer partyDelNy) {
 		this.partyDelNy = partyDelNy;
 	}
-
 	public Date getPartyRegDate() {
 		return partyRegDate;
 	}
-
 	public void setPartyRegDate(Date partyRegDate) {
 		this.partyRegDate = partyRegDate;
 	}
-
 	public Date getPartyModDate() {
 		return partyModDate;
 	}
-
 	public void setPartyModDate(Date partyModDate) {
 		this.partyModDate = partyModDate;
 	}
-
 	public String getxMasterSeqF() {
 		return xMasterSeqF;
 	}
-
 	public void setxMasterSeqF(String xMasterSeqF) {
 		this.xMasterSeqF = xMasterSeqF;
 	}
-
 	public String getxMemberName() {
 		return xMemberName;
 	}
-
 	public void setxMemberName(String xMemberName) {
 		this.xMemberName = xMemberName;
 	}
-
 	public String getxPartySeqF() {
 		return xPartySeqF;
 	}
-
 	public void setxPartySeqF(String xPartySeqF) {
 		this.xPartySeqF = xPartySeqF;
 	}
-
 	public String getxMemberSeqF() {
 		return xMemberSeqF;
 	}
-
 	public void setxMemberSeqF(String xMemberSeqF) {
 		this.xMemberSeqF = xMemberSeqF;
 	}
-
 	public String getxMasterNy() {
 		return xMasterNy;
 	}
-
 	public void setxMasterNy(String xMasterNy) {
 		this.xMasterNy = xMasterNy;
 	}
-
 	public String getxPartyJoinNy() {
 		return xPartyJoinNy;
 	}
-
 	public void setxPartyJoinNy(String xPartyJoinNy) {
 		this.xPartyJoinNy = xPartyJoinNy;
 	}
-
 	public Integer getxMemberTierCd() {
 		return xMemberTierCd;
 	}
-
 	public void setxMemberTierCd(Integer xMemberTierCd) {
 		this.xMemberTierCd = xMemberTierCd;
 	}
-
-	public Integer getxMemberVictoryRate() {
+	public Double getxMemberVictoryRate() {
 		return xMemberVictoryRate;
 	}
-
-	public void setxMemberVictoryRate(Integer xMemberVictoryRate) {
+	public void setxMemberVictoryRate(Double xMemberVictoryRate) {
 		this.xMemberVictoryRate = xMemberVictoryRate;
 	}
-
-	public double getxMemberKDA() {
+	public Double getxMemberKDA() {
 		return xMemberKDA;
 	}
-
-	public void setxMemberKDA(double xMemberKDA) {
+	public void setxMemberKDA(Double xMemberKDA) {
 		this.xMemberKDA = xMemberKDA;
 	}
-
 	public Date getxPartyJoinDate() {
 		return xPartyJoinDate;
 	}
-
 	public void setxPartyJoinDate(Date xPartyJoinDate) {
 		this.xPartyJoinDate = xPartyJoinDate;
 	}
+
+	
 
 }
